+++ R/graph.curve.params.R
@@ -8,32 +8,31 @@
 #'
 #' @examples
 #' \dontrun{
-#' curve_params <- readRDS(url("https://osf.io/download/rtw5k/"))
-#'
 #' plot1 <- graph.curve.params(curve_params)
 #'
 #' print(plot1)
 #' }
-graph.curve.params = function(
+graph.curve.params <- function(
     curve_params,
     antigen_isos = unique(curve_params$antigen_iso),
-    verbose = FALSE)
-{
-
-  if (verbose)
+    verbose = FALSE) {
+  if (verbose) {
     message(
       "Graphing curves for antigen isotypes: ",
-      paste(antigen_isos, collapse = ", "))
+      paste(antigen_isos, collapse = ", ")
+    )
+  }
 
-  curve_params = curve_params %>%
+  curve_params <- curve_params %>%
     dplyr::filter(.data$antigen_iso %in% antigen_isos)
 
   day2yr <- 365.25
 
-  tx2 <- 10 ^ seq(-1, 3.1, 0.025)
+  tx2 <- 10^seq(-1, 3.1, 0.025)
 
-  bt <- function(y0, y1, t1)
+  bt <- function(y0, y1, t1) {
     log(y1 / y0) / t1
+  }
 
   # uses r > 1 scale for shape
   ab <- function(t, y0, y1, t1, alpha, shape) {
@@ -41,14 +40,15 @@ graph.curve.params = function(
 
     yt <- 0
 
-    if (t <= t1)
+    if (t <= t1) {
       yt <- y0 * exp(beta * t)
+    }
 
-    if (t > t1)
-      yt <- (y1 ^ (1 - shape) - (1 - shape) * alpha * (t - t1)) ^ (1 / (1 - shape))
+    if (t > t1) {
+      yt <- (y1^(1 - shape) - (1 - shape) * alpha * (t - t1))^(1 / (1 - shape))
+    }
 
     return(yt)
-
   }
 
 
@@ -68,15 +68,18 @@ graph.curve.params = function(
     ) %>%
     dplyr::slice(
       rep(1:dplyr::n(),
-          each = nrow(d)))
+        each = nrow(d)
+      )
+    )
 
 
   serocourse.all <-
-    cbind(d, dT)  %>%
+    cbind(d, dT) %>%
     tidyr::pivot_longer(
       cols = dplyr::starts_with("time"),
-      values_to = "t") %>%
-    select(-"name")  %>%
+      values_to = "t"
+    ) %>%
+    select(-"name") %>%
     rowwise() %>%
     mutate(res = ab(
       .data$t,
@@ -84,20 +87,21 @@ graph.curve.params = function(
       .data$y1,
       .data$t1,
       .data$alpha,
-      .data$r)) %>%
+      .data$r
+    )) %>%
     ungroup()
 
-  if (verbose) message('starting to compute quantiles')
+  if (verbose) message("starting to compute quantiles")
   serocourse.sum <- serocourse.all %>%
     summarise(
       .by = c("antigen_iso", "t"),
-      res.med  = quantile(.data$res, 0.5),
-      res.low  = quantile(.data$res, 0.025),
+      res.med = quantile(.data$res, 0.5),
+      res.low = quantile(.data$res, 0.025),
       res.high = quantile(.data$res, 0.975),
-      res.p75  = quantile(.data$res, 0.75),
-      res.p25  = quantile(.data$res, 0.25),
-      res.p10  = quantile(.data$res, 0.10),
-      res.p90  = quantile(.data$res, 0.90)
+      res.p75 = quantile(.data$res, 0.75),
+      res.p25 = quantile(.data$res, 0.25),
+      res.p10 = quantile(.data$res, 0.10),
+      res.p90 = quantile(.data$res, 0.90)
     ) %>%
     pivot_longer(
       names_to = "quantile",
@@ -119,39 +123,42 @@ graph.curve.params = function(
   ggplot2::ggplot() +
     ggplot2::geom_line(
       data = serocourse.sum %>%
-        filter(.data$quantile == "med") ,
+        filter(.data$quantile == "med"),
       ggplot2::aes(
         x = .data$t,
-        y = .data$res),
+        y = .data$res
+      ),
       linewidth = 1
     ) +
     ggplot2::geom_line(
-      data = serocourse.sum %>% filter(quantile == "p10") ,
+      data = serocourse.sum %>% filter(quantile == "p10"),
       ggplot2::aes(
         x = .data$t,
-        y = .data$res),
+        y = .data$res
+      ),
       linewidth = .5
     ) +
     ggplot2::geom_line(
       data = serocourse.sum %>%
-        filter(quantile == "p90") ,
+        filter(quantile == "p90"),
       ggplot2::aes(
         x = .data$t,
-        y = .data$res),
+        y = .data$res
+      ),
       linewidth = .5
     ) +
     ggplot2::facet_wrap(
       ~ .data$antigen_iso,
-      ncol = 2) +
+      ncol = 2
+    ) +
     ggplot2::scale_y_log10(
       limits = c(0.9, 2000),
       breaks = c(1, 10, 100, 1000),
       minor_breaks = NULL
     ) +
-    ggplot2::theme_minimal()  +
+    ggplot2::theme_minimal() +
     ggplot2::theme(axis.line = ggplot2::element_line()) +
     ggplot2::labs(x = "Days since fever onset", y = "ELISA units")
-
 }
 
 # ggplot() +