+++ R/load_curve_params.R
@@ -10,34 +10,15 @@
 #'
 #' print(curve)
 #'
-#'
-load_curve_params = function(file_path, antigen_isos = NULL)
-{
-  if(file_path %>% substr(1,4) == "http")
-  {
-    file_path = url(file_path)
-
+load_curve_params <- function(file_path, antigen_isos = NULL) {
+  if (file_path %>% substr(1, 4) == "http") {
+    file_path <- url(file_path)
   }
 
-  curve_params =
+  curve_params <-
     file_path %>%
     readRDS() %>%
-    tibble::as_tibble()
-
-  class(curve_params) =
-    c("curve_params", class(curve_params))
-
-  if(is.null(antigen_isos))
-  {
-    antigen_isos = unique(curve_params$antigen_iso)
-  } else
-  {
-    stopifnot(all(is.element(antigen_isos, curve_params$antigen_iso)))
-
-  }
-
-  attr(curve_params, "antigen_isos") = antigen_isos
+    as_curve_params()
 
   return(curve_params)
-
 }